requirements/base.txt
# 移动原 requirements.txt 的内容到这里 
aiohappyeyeballs==2.4.3
aiohttp==3.11.3
aiosignal==1.3.1
alibabacloud-domain20180129==4.0.0
alibabacloud-ecs20140526==4.4.7
alibabacloud-tea==0.4.0
alibabacloud_credentials==0.3.6
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.2
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.12
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.3
django-apscheduler==0.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
PyJWT==2.10.0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.17.2
